+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # :arrows_counterclockwise: NetApp ONTAP Storage As Built Report Changelog
 
+## [0.6.5] - 2022-11-06
+
+### Added
+
+- Added aggregate spare reporting [#26](https://github.com/AsBuiltReport/AsBuiltReport.NetApp.ONTAP/issues/26)
+- Added Ontap Multi Admin Approval [#29](https://github.com/AsBuiltReport/AsBuiltReport.NetApp.ONTAP/issues/29)
+- Added Consistency Group Support [#28](https://github.com/AsBuiltReport/AsBuiltReport.NetApp.ONTAP/issues/28)
+- Added Audit Logs Support [#31](https://github.com/AsBuiltReport/AsBuiltReport.NetApp.ONTAP/issues/31)
+- Added Audit log destination [#30](https://github.com/AsBuiltReport/AsBuiltReport.NetApp.ONTAP/issues/30)
+
+### Fixed
+
+- Fix #22
+- Fix #23
+- Fix #24
+- Close #29
+- Close #26
+- Close #28
+- Close #30
+- Close #31
+- Fixed SNMP section not shown in report.
+
 ## [0.6.4] - 2022-05-14
 
 ### Added